Support for older Android devices

+++ app/src/main/java/com/example/batteryblynkstatus/BatteryStatus.kt
@@ -1,3 +1,3 @@
 package com.example.batteryblynkstatus
 
-class BatteryStatus(val level: Int, val allCapacity: Double, val currentCapacity: Double)
+class BatteryStatus(val level: Int, val allCapacity: Int, val currentCapacity: Int)
